Fix half send reqs issue
If we set set:get ratio to 0:1, memtier will send out exactly 1 get request and
increase the get command counter to 1, then when we issue the second request, it
will just set the counter back to 0 and do nothing. It's a design issue of
memtier but it is ok for it. Because it will try to fill the pipeline with no
inter-arrival time. So it will just try multiple times to issue the request.

This commit fix the problem by checking at the end of create_request() every
time.