[WIP] Fix distribution generator issue, only clip based on pipeline size
Remove the clipping condition that uses nextCycle. Try to send as fast as
possible if we cannot meet the request rate and limit the outstanding requests
by the pipeline size.

Also, write to socket when its writable and don't wait when its not writable.
Check the write buffer later instead of busy waiting.